Scheduled operations
Split scheduled operation objects into one class pr file. Added more
comments to classes.
