+++ RsyncOSXver30/Resources.swift
@@ -20,7 +20,7 @@ struct Resources {
     // Resource strings
     private var changelog:String = "https://github.com/rsyncOSX/Documentation/blob/master/docs/Changelog.md"
     private var documents: String = "https://github.com/rsyncOSX/Documentation"
-    private var urlPlist: String = "https://dl.dropboxusercontent.com/u/52503631/versionRsyncOSX.plist?raw=1"
+    private var urlPlist: String = "https://raw.githubusercontent.com/rsyncOSX/Version3.x/master/versionRsyncOSX/versionRsyncOSX.plist"
     
     // Get the resource.
     func getResource (resource: resourceType) -> String {
